refactor: move message reporting to plugin
