+++ gravitee-apim-gateway/gravitee-apim-gateway-core/src/main/java/io/gravitee/gateway/reactive/core/analytics/AnalyticsContext.java
@@ -1,0 +1,55 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.gateway.reactive.core.analytics;
+
+import io.gravitee.definition.model.v4.analytics.Analytics;
+import lombok.Getter;
+import lombok.extern.slf4j.Slf4j;
+
+/**
+ * @author Guillaume LAMIRAND (guillaume.lamirand at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+@Getter
+@Slf4j
+public class AnalyticsContext {
+
+    protected Analytics analytics;
+    private LoggingContext loggingContext;
+
+    public AnalyticsContext(final Analytics analytics, final String loggingMaxsize, final String loggingExcludedResponseType) {
+        this.analytics = analytics;
+        if (analytics != null && analytics.isEnabled()) {
+            initLoggingContext(loggingMaxsize, loggingExcludedResponseType);
+        }
+    }
+
+    private void initLoggingContext(final String loggingMaxsize, final String loggingExcludedResponseType) {
+        if (AnalyticsUtils.isLoggingEnabled(analytics)) {
+            this.loggingContext = new LoggingContext(analytics.getLogging());
+            this.loggingContext.setMaxSizeLogMessage(loggingMaxsize);
+            this.loggingContext.setExcludedResponseTypes(loggingExcludedResponseType);
+        }
+    }
+
+    public boolean isEnabled() {
+        return this.analytics.isEnabled();
+    }
+
+    public boolean isLoggingEnabled() {
+        return this.loggingContext != null;
+    }
+}
