+++ gravitee-apim-gateway/gravitee-apim-gateway-handlers/gravitee-apim-gateway-handlers-api/src/main/java/io/gravitee/gateway/reactive/handlers/api/analytics/logging/LoggingHook.java
@@ -20,10 +20,10 @@ import io.gravitee.gateway.reactive.api.ExecutionPhase;
 import io.gravitee.gateway.reactive.api.context.ExecutionContext;
 import io.gravitee.gateway.reactive.api.context.InternalContextAttributes;
 import io.gravitee.gateway.reactive.api.hook.InvokerHook;
+import io.gravitee.gateway.reactive.core.analytics.AnalyticsContext;
+import io.gravitee.gateway.reactive.core.analytics.LoggingContext;
 import io.gravitee.gateway.reactive.core.context.MutableExecutionContext;
 import io.gravitee.gateway.reactive.core.context.MutableResponse;
-import io.gravitee.gateway.reactive.core.v4.analytics.AnalyticsContext;
-import io.gravitee.gateway.reactive.core.v4.analytics.LoggingContext;
 import io.gravitee.gateway.reactive.handlers.api.analytics.logging.request.LogEndpointRequest;
 import io.gravitee.gateway.reactive.handlers.api.analytics.logging.response.LogEndpointResponse;
 import io.gravitee.reporter.api.v4.log.Log;
