+++ gravitee-apim-gateway/gravitee-apim-gateway-handlers/gravitee-apim-gateway-handlers-api/src/main/java/io/gravitee/gateway/reactive/handlers/api/analytics/logging/request/LogEndpointRequest.java
@@ -18,7 +18,7 @@ package io.gravitee.gateway.reactive.handlers.api.analytics.logging.request;
 import io.gravitee.gateway.api.http.HttpHeaders;
 import io.gravitee.gateway.reactive.api.context.ExecutionContext;
 import io.gravitee.gateway.reactive.api.context.Request;
-import io.gravitee.gateway.reactive.core.v4.analytics.LoggingContext;
+import io.gravitee.gateway.reactive.core.analytics.LoggingContext;
 import io.gravitee.reporter.api.v4.metric.Metrics;
 
 /**
