+++ gravitee-apim-plugin/gravitee-apim-plugin-reactor/gravitee-apim-plugin-reactor-v4/src/main/java/io/gravitee/apim/plugin/reactor/core/analytics/sampling/ProbabilityMessageStrategy.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.gateway.reactive.core.v4.analytics.sampling;
+package io.gravitee.apim.plugin.reactor.core.analytics.sampling;
 
 import io.gravitee.gateway.reactive.api.message.Message;
+import java.util.SplittableRandom;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
@@ -26,31 +27,36 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Slf4j
 @Getter(AccessLevel.PROTECTED)
-public class CountMessageSamplingStrategy implements MessageSamplingStrategy {
+public class ProbabilityMessageStrategy implements MessageSamplingStrategy {
 
-    private static final int DEFAULT_COUNT = 100;
-    private static final int DEFAULT_MIN_COUNT = 10;
-    private int count;
+    private static final Double DEFAULT_PROBABILITY = 0.01;
+    private static final Double DEFAULT_MAX_PROBABILITY = 0.5;
+    private final SplittableRandom random = new SplittableRandom();
+    private double probability;
 
-    public CountMessageSamplingStrategy(final String value) {
+    public ProbabilityMessageStrategy(final String value) {
         try {
-            this.count = Integer.parseInt(value);
-            if (this.count < DEFAULT_MIN_COUNT) {
+            this.probability = Double.parseDouble(value);
+            if (this.probability > DEFAULT_MAX_PROBABILITY) {
                 log.warn(
-                    "Probability sampling value '{}' is lower than minimum allowed '{}', using min value instead.",
+                    "Probability sampling value '{}' is higher than maximum allowed '{}', using max value instead.",
                     value,
-                    DEFAULT_MIN_COUNT
+                    DEFAULT_MAX_PROBABILITY
                 );
-                this.count = DEFAULT_MIN_COUNT;
+                this.probability = DEFAULT_MAX_PROBABILITY;
             }
         } catch (Exception e) {
-            log.warn("Count sampling value '{}' cannot be parsed as Integer, using default value '{}'", value, DEFAULT_COUNT);
-            this.count = DEFAULT_COUNT;
+            log.warn("Probability sampling value '{}' cannot be parsed as Double, using default value '{}'", value, DEFAULT_PROBABILITY);
+            this.probability = DEFAULT_PROBABILITY;
         }
     }
 
     @Override
     public boolean isRecordable(final Message message, final int messageCount, final long lastMessageTimestamp) {
-        return messageCount == 1 || messageCount % count == 1;
+        return messageCount == 1 || matchesProbability(random.nextDouble());
+    }
+
+    protected boolean matchesProbability(final Double random) {
+        return random < probability;
     }
 }
