+++ gravitee-apim-plugin/gravitee-apim-plugin-reactor/gravitee-apim-plugin-reactor-v4/src/main/java/io/gravitee/apim/plugin/reactor/handlers/api/analytics/logging/MessageLog.java
@@ -16,7 +16,7 @@
 package io.gravitee.apim.plugin.reactor.handlers.api.analytics.logging;
 
 import io.gravitee.gateway.api.http.HttpHeaders;
-import io.gravitee.gateway.reactive.core.v4.analytics.LoggingContext;
+import io.gravitee.gateway.reactive.core.analytics.LoggingContext;
 import io.gravitee.reporter.api.v4.common.Message;
 import java.util.HashMap;
