+++ gravitee-apim-plugin/gravitee-apim-plugin-reactor/gravitee-apim-plugin-reactor-v4/src/main/java/io/gravitee/apim/plugin/reactor/processor/message/reporter/EntrypointRequestReporterMessageProcessor.java
@@ -15,16 +15,16 @@
  */
 package io.gravitee.apim.plugin.reactor.processor.message.reporter;
 
+import io.gravitee.apim.plugin.reactor.core.analytics.V4AnalyticsContext;
+import io.gravitee.apim.plugin.reactor.core.analytics.sampling.MessageSamplingStrategy;
 import io.gravitee.apim.plugin.reactor.handlers.api.analytics.MessageAnalyticsHelper;
 import io.gravitee.apim.plugin.reactor.handlers.api.analytics.MessageCounters;
 import io.gravitee.apim.plugin.reactor.handlers.api.analytics.MessageReporter;
 import io.gravitee.apim.plugin.reactor.handlers.api.analytics.logging.request.message.MessageLogEntrypointRequest;
 import io.gravitee.gateway.reactive.api.context.InternalContextAttributes;
+import io.gravitee.gateway.reactive.core.analytics.LoggingContext;
 import io.gravitee.gateway.reactive.core.context.MutableExecutionContext;
 import io.gravitee.gateway.reactive.core.processor.MessageProcessor;
-import io.gravitee.gateway.reactive.core.v4.analytics.AnalyticsContext;
-import io.gravitee.gateway.reactive.core.v4.analytics.LoggingContext;
-import io.gravitee.gateway.reactive.core.v4.analytics.sampling.MessageSamplingStrategy;
 import io.gravitee.gateway.report.ReporterService;
 import io.gravitee.reporter.api.v4.common.MessageConnectorType;
 import io.gravitee.reporter.api.v4.common.MessageOperation;
@@ -58,7 +58,7 @@ public class EntrypointRequestReporterMessageProcessor implements MessageProcess
             return Completable.complete();
         }
         return Completable.defer(() -> {
-            final AnalyticsContext analyticsContext = ctx.getInternalAttribute(InternalContextAttributes.ATTR_INTERNAL_ANALYTICS_CONTEXT);
+            final V4AnalyticsContext analyticsContext = ctx.getInternalAttribute(InternalContextAttributes.ATTR_INTERNAL_ANALYTICS_CONTEXT);
 
             if (analyticsContext == null || !analyticsContext.isEnabled()) {
                 return Completable.complete();
