+++ gravitee-apim-plugin/gravitee-apim-plugin-reactor/gravitee-apim-plugin-reactor-v4/src/main/java/io/gravitee/apim/plugin/reactor/processor/message/reporter/EntrypointResponseReporterMessageProcessor.java
@@ -20,10 +20,10 @@ import io.gravitee.apim.plugin.reactor.handlers.api.analytics.MessageCounters;
 import io.gravitee.apim.plugin.reactor.handlers.api.analytics.MessageReporter;
 import io.gravitee.apim.plugin.reactor.handlers.api.analytics.logging.response.message.MessageLogEntrypointResponse;
 import io.gravitee.gateway.reactive.api.context.InternalContextAttributes;
+import io.gravitee.gateway.reactive.core.analytics.AnalyticsContext;
+import io.gravitee.gateway.reactive.core.analytics.LoggingContext;
 import io.gravitee.gateway.reactive.core.context.MutableExecutionContext;
 import io.gravitee.gateway.reactive.core.processor.MessageProcessor;
-import io.gravitee.gateway.reactive.core.v4.analytics.AnalyticsContext;
-import io.gravitee.gateway.reactive.core.v4.analytics.LoggingContext;
 import io.gravitee.gateway.report.ReporterService;
 import io.gravitee.reporter.api.v4.common.MessageConnectorType;
 import io.gravitee.reporter.api.v4.common.MessageOperation;
