+++ gravitee-apim-plugin/gravitee-apim-plugin-reactor/gravitee-apim-plugin-reactor-v4/src/test/java/io/gravitee/apim/plugin/reactor/core/analytics/sampling/CountMessageSamplingStrategyTest.java
@@ -13,16 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.gateway.reactive.core.v4.analytics.sampling;
+package io.gravitee.apim.plugin.reactor.core.analytics.sampling;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import io.gravitee.definition.model.v4.analytics.sampling.Sampling;
-import io.gravitee.definition.model.v4.analytics.sampling.SamplingType;
 import io.gravitee.gateway.reactive.api.message.DefaultMessage;
-import java.time.Duration;
-import java.time.temporal.ChronoUnit;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
