+++ gravitee-apim-plugin/gravitee-apim-plugin-reactor/gravitee-apim-plugin-reactor-v4/src/test/java/io/gravitee/apim/plugin/reactor/handlers/api/analytics/logging/response/message/MessageLogEndpointResponseTest.java
@@ -20,7 +20,7 @@ import static org.mockito.Mockito.when;
 
 import io.gravitee.gateway.api.http.HttpHeaders;
 import io.gravitee.gateway.reactive.api.message.DefaultMessage;
-import io.gravitee.gateway.reactive.core.v4.analytics.LoggingContext;
+import io.gravitee.gateway.reactive.core.analytics.LoggingContext;
 import java.util.HashMap;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
